+ make: convert tox/poetry to make/pip-tools/setuptools/build
new maintainer, new tools preference... make and pip-tools are more
familiar for me.

this commit also updates everything to PEP518 and PEP621 build and
metadata.  we have chosen not to directly embed requirements in
pyproject.toml however, even though pip-tools can read it, because it
seems more appropriate to track these in their own files.  the files are
still referenced in pyproject.toml.

also included are the serialized package metadata generated by the
pip-compile run (.txt files), as this is the way to fully reproduce a
tested build, by keeping track of these data in scm.  otherwise build
calculations can change the resulting environment, even though the .in
files haven't changed.

we have also set a new release minor version to signify this larger
change, and the new maintainer.

docs are not yet moved over because we will change how those are
implemented, however tests do work (besides the test errors due to timew
output changing, which we will fix in an upcoming commit).