/ tests: timewarrior: fix failing tag tests, had extraneous quotes
not sure why these had quote chars originally but they make the tests fail.
the quotes are not needed because they are already made into separate words for
preparation of the eventual shell command words:

  $ python3 -q
  >>> from timew import TimeWarrior
  >>> tw = TimeWarrior(simulate=True)
  >>> tw.tag(1, ["new tag", "another tag"])
  ['timew', 'tag', '@1', 'new tag', 'another tag']

don't really like that these tags have spaces, but then again it's good
to test that they do in fact work and have expected output.  so we'll
just fix the quote issue for now

once this fix is in place, all the tests work again