+ pyproject: authors: add myself so my email shows up on pypi
makes sense for my name to show up on pypi, in case someone wants to
email about it, given my responsibility as maintainer.  really though,
my contribution is minimal compared to the original author.  not sure if
he wants his name there, because then he could get emails related to it,
but nonetheless he needs to be credited as the principal author.