- timew: remove side effects using None instead of now() in kwargs
this is effectively the same, because using an open-ended interval ends
now anyways.

it has the added benefit of not changing apidocs render via the side
effect of running datetime.now() on import and thus changing the docs
every time they are built.  should really not have side effects in a
kwarg default anyways...