+ make: pybase: change python to 3.12, docs build broke somehow
not sure why, but the python3.9 stopped working for some reason, not
clear to me why since it worked before.  but now tried it with 3.12 and
docs build works.  so we will bump the python needed to build it, but we
will leave the python needed for downstream consumers to be at 3.9
still.  not the greatest where the tested version will be later than the
minimum supported version, but there's no "> 3.9"-isms in the codebase
(and it's small enough) not to worry about this yet